Fix for non-idempotent DELETE on documents
Not yet implemented for folders; unsure if this behavior applies to them.
Basically, a DELETE request against a document in the trash permanently
deletes it. This is not expected behavior; DELETE requests should be
idempotent. #delete! has been added to allow you to still permanently delete
objects, should that be necessary.